Remind users to have correct FASTA deflines
VirSorter replaces some special characters (%!?|, etc.) with the _ character. These special characters shouldn't exist in FASTA deflines anyways if the file is going to be used with anvi'o, but I'm adding a reminder just to be sure. If any special characters sneak through and enter the VirSorter pipeline, the VirSorterParser won't work because contigs will have different names in VirSorter results or anvi'o.