suggestions for conda install instructions
Removing this part: 

>Then, add the following channels that will be required for the anvi'o installation:
>
>``` bash
>conda config --env --add channels conda-forge
>conda config --env --add channels bioconda
>```

Because that changes the user's base conda environment (which i learned myself recently, but makes sense in hindsight), which we don't need to do, as we can just provide the channels in the install command. Listing the channels in the install command (as changed above) will also help with some people apparently missing the first part about setting channels, who are getting stuck at the subsequent "package not found" error (which can then be confusing because another possible cause is listed as a warning at that stage).