don't use 'text=""' for PROGRAMs and ARTIFACTs if it is not necesary.
If there is no text variable, the template inserts the artifact/program
name as text. I was on the fence whether it is worth to remove all, but
then decided that it is better to have consistency that reflects the
default and special cases for future technical needs. Ther is no need to
change anything outside of this tutorial that are done, but perhaps it
is something to keep in mind for future references :)